Don't refresh variables on every console execution (#5871)
This change addresses an issue in which invoking long-running commands
causes variables RPC errors to appear.

The underlying issue is that currently every code execution in the
Console is triggering a full refresh of the Variables pane (introduced
in https://github.com/posit-dev/positron/pull/5583). So when code is
executed, we're asking for a refreshed Variables pane immediately
afterwards, but the RPC doesn't go through right away because code is
executing. If the code takes a while to execute, the RPC eventually
times out.

The fix is to avoid refreshing the Variables pane if it's already
showing the correct session.

There are other changes that would improve this, such as having the
Variables pane queue refresh requests until the kernel is idle, or
auditing refresh triggers so they run when code _finishes_ executing.
This change is intended to be the minimal one that fixes the problem.

Addresses https://github.com/posit-dev/positron/issues/5813 and
https://github.com/posit-dev/positron/issues/5806.

### QA Notes

Make sure the behavior described in
https://github.com/posit-dev/positron/pull/5583 still works, too.