Data Explorer: Support opening gzipped csv files with DuckDB by decompressing in memory in NodeJS (#5901)
Addresses #5332. This works around duckdb-wasm's limitation that it
cannot read directly from gzipped CSV/TSV files by unzipping in memory
and registering as a virtual file. This is okay for small files but for
very large gzipped files will feel a bit slow and use a bunch of memory
(we will have to tackle that problem via #5889).

### Release Notes

#### New Features

- Support opening gzipped CSV/TSV files in the data explorer with .gz
extension from the file explorer pane (#5332).

#### Bug Fixes

- N/A

e2e: @:data-explorer