Add strip def to rpm spec template (#5913)
This change makes it possible to use a custom strip binary for Positron
RPM builds. The upstream copy of the template already has this line (so
this is just updating ours to match it).


https://github.com/posit-dev/positron/blob/be85387931ff23fc51a080a28998adeea64f59be/resources/linux/rpm/code.spec.template#L19-L22

This change is in service of the larger effort to make arm64 builds of
Positron possible.