Don't start runtimes that were never used (#5992)
This change cause Positron to "forget" runtimes that are affiliated with
the workspace but that aren't actually used, with the goal of preventing
unwanted sessions from starting.

Some details:

- A runtime is considered to be unused if it was auto-started as an
affiliated runtime but then never actually used in the last session.
- The runtime affiliation is not wholly removed; we just stop
auto-starting it when Positron starts. So e.g. it is still consulted
when auto-starting a console in order to run code.
- This only applies to projects with multiple runtime affiliations. If
you only have one runtime affiliated, it will auto-start even if you
didn't use it in your last session.

Addresses #3432.

### Release Notes

#### New Features

- Do not auto-start interpreters that aren't being used in the workspace
(#3432)

#### Bug Fixes

- N/A


### QA Notes

For the purposes of this feature, a runtime is considered to have been
"used" if you switch to it in the Console.