Fix `lightning` and `torch._dynamo` (#5984)
Addresses #5879. I confirmed that the test fails before pinning
`typing-extensions`.

I haven't dug further into why this happens or how to fix it at the
level of `typing-extensions`. We may need to do that in future if we
encounter this again.

Note that this PR also updated a few other vendored packages. The
`pygments` update required a tiny change to one of our Help-related
tests.

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Fixed a recent regression causing an error when trying to import
`torch._dynamo`, and therefore `lightning` (#5879).

### QA Notes

Try to `import lightning` in a Python console. Note that you will have
to rerun `npm install` in the `positron-python` extension to install the
new vendored package versions.