Pull in missing 1.96 upstream changes (#6024)
Upstream `main` was missing a few changes that are needed for the
Workbench patch, see PR:
https://github.com/rstudio/vscode-server/pull/162

- address Snyk vulnerabilities in micromatch and dompurify
- use correct table name when saving state to Workbench
- update extension host agent server to send first heartbeat on startup
in Workbench
- bump workbench extension

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->

---------

Co-authored-by: petetronic <1178561+petetronic@users.noreply.github.com>
Co-authored-by: Jonathan McPherson <jonathan@posit.co>