e2e test: Duplicate variables entries (#6022)
### Intent 

To help test #5887 these new tests ensure that after an interpreter env is restarted, it doens't create a duplicate entry in variables pane.

### Approach

* Restarts interpreter and checks that only 1 still exists
* R & Python.. I thought adding a notebook version would be unnecessary, but can if needed.

### QA Notes

Tests are skipped now due to the issue. But you can check the tests by removing the skip and it will fail with

```
    Error: expect(received).toBe(expected) // Object.is equality

    Expected: 1
    Received: 2

```
All other variable tests should still pass 
@:variables

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->