Fix issue preventing drag-and-drop files into Explorer view (#6030)
This change addresses an issue that prevents file drag-n-drop from
working in the Explorer view.

The problem was that there was no Electron implementation of the
accessors that Posit Workbench uses to check for file upload/download
permissions. Consequently, in desktop mode these accessors did not work.

The fix is to add accessors to the native (Electron) version of the
environment service.

Addresses https://github.com/posit-dev/positron/issues/5266.


### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Fix file drag and drop into the Explorer view (#5266) 

### QA Notes

This change doesn't touch the logic for Workbench so testing desktop
should be sufficient.