Fix behavior when missing Pip (#6020)
When selecting an interpreter missing ipykernel, the UI prompt now
states that both Pip and ipykernel must be installed. `ProductInstaller`
logic was modified to pass flags to enable Pip to be installed
automatically. Additional flags were set in `ProductInstaller` to enable
the Pip install to respect the `python.installModulesInTerminal`
setting.

Addresses #5505