Don't veto shutdown in web mode (#6091)
This change fixes an issue in which Positron prompts to save changes
whenever you try to navigate away from the page in web mode.

It turns out that in web mode (specifically), async shutdown operations
aren't supported, and trying to use them results in this behavior. The
fix is simply to not clear the session list in web mode, which should be
fine as the reason we need to do this has to do with multiple Electron
windows.

Addresses https://github.com/posit-dev/positron/issues/6062. 

### QA Notes

As noted in the issue it's very easy to hit this, so should be easy to
verify as well!