Be much more gracious with supervisor startup time (#6100)
This change addresses an issue frequently seen in tests wherein the
kernel supervisor appears to fail to connect at startup. After several
rounds of debugging, it eventually became clear that there was no actual
reconnection failure; the problem was simply that the CI machines
(especially on Windows hardware) are very slow, and the client gave up
retrying before the supervisor process was fully started.

The fix is wait much longer before giving up, and to base the number of
retries on wall clock time rather than attempts. Formerly, we could
exhaust retries in as little as 1.5-2 seconds; now we wait up to 10
seconds.

Here are 2 e2e test runs from a branch with this change:

https://github.com/posit-dev/positron/actions/runs/12934857967
https://github.com/posit-dev/positron/actions/runs/12934185231

Addresses #5337.

### QA Notes

This change is primarily intended to address the issue in CI and
shouldn't have much impact on the product; the only downside of this
change is that now if there really _is_ some issue that causes a
connectivity failure, it takes 10 seconds for it to show up instead of
2.

These situations should not be common in the wild, since practically all
of the issues we have seen manifest as an error starting/launching the
supervisor, which we can detect without waiting for a timeout.