Bump bundled Quarto version (#6103)
Addresses #5643 by updating to the [current release
version](https://quarto.org/docs/download/) (not the pre-release).

Given that our tests don't run against the release build, I don't know
that we have anything to gain by running any specific tests here. I
_will_ kick off a release build for some manual spot checking.

We waited to do this because we do _not_ want this to go into this
month's Workbench patch release (Quarto's new Deno base), but we do
hope/expect this to land in the 2025.04 Workbench release.

### Release Notes

#### New Features

- Updated bundled Quarto to 1.6.40, the current release version

#### Bug Fixes

- N/A


### QA Notes

When this is in a release build, we should confirm that Quarto workflows
are all good to go!