python: add Python 3.13 testing in CI (#6108)
In an effort to get testing available for Python 3.13, let's skip the
tests for ibis, since it cannot be downloaded for 3.13 anyway.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- addresses #5137 by adding in testing infrastructure for Python 3.13

#### Bug Fixes

- N/A


### QA Notes

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->

should pass CI

---------

Signed-off-by: Isabel Zimmerman <54685329+isabelizimm@users.noreply.github.com>