Speculative fix for AggregateError at startup (#6130)
This change is a speculative fix for an error that has been observed
when starting R sessions.

The problem (based on analysis of the logs) results from the new logic
the supervisor has to clean itself up if it is left idle for too long,
an improvement added to help support #4996. What appears to be happening
is that in some cases, the supervisor is performing an idle check the
instant it starts up, then determining itself to be idle (which of
course it _is_), then exiting.

There are two changes here:

- First, use a new build of the supervisor that ensures that the first
idle check doesn't happen prematurely (0.28 => 0.29)
- Second, the "AggregateError" message is not very helpful, so added
logic to help unpack AggregateErrors when summarizing errors thrown
across the API boundary.

Addresses https://github.com/posit-dev/positron/issues/6129.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Speculative fix for occasional `AggregateError` seen when starting a
new console (#6129).


### QA Notes

Unfortunately we don't have a repro for this.