Add command for opening an HTML file in a browser (#6071)
This change adds an "Open in Browser" command to HTML files to go with
"Open in Viewer", based on user feedback that it is not obvious how to
open an HTML file in a browser if you're looking at it in Positron.

Addresses #6047. 

### Release Notes

#### New Features

- New "Open in Browser" command for HTML files (#6047)

#### Bug Fixes

- N/A


### QA Notes

- This change exists only in the Positron Proxy extension and thus
should not affect any core scenarios.
- I changed the icon for the "Open in Viewer" command to "preview" since
I think it's better suited to the gesture.