Display active runtime metadata in console pane (#6156)
Addresses #5999 

Adds a new feature flag called `positron.multipleConsoleSessions`. When
the feature flag is on users can now view details for an active session
by clicking on the new info icon button that lives in the console pane
action bar. A modal popup will render, containing the details of the
session, such as name, state, id, interpreter path, and environment
management tool.

The popup provides a link to the session logs in the output pane. When
clicked, the user is taken to the output pane and shown the kernel
channel for their interpreter.

This is a first pass for the info panel - additional information may be
added based off feedback! Follow up work to add additional links to
other channels is covered in #6149.

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes
- [ ] Verify the info button is only shown when the feature flag is on
and there is an active session for the console instance
- Note: An issue has been noted where the popup is not rendered if the
console pane height is smaller than the height of the popup (#3061)
- [ ] Verify the session status shown in the popup updates in real time

### Screenshots

**Feature Flag**
<img width="1272" alt="Screenshot 2025-01-29 at 2 19 30 PM"
src="https://github.com/user-attachments/assets/84154034-f7a7-4696-b016-61c4aa3c9370"
/>

**Popup**
<img width="1330" alt="Screenshot 2025-01-29 at 4 38 16 PM"
src="https://github.com/user-attachments/assets/fc819ef6-ad75-4e59-9217-47b97c965b41"
/>

**Real-Time Session State Updates**

https://github.com/user-attachments/assets/16b0b81f-7b48-43d0-9f54-047e6e2af4be