Add new R extension quickpick to select an interpreter (#6213)
Addresses #5223

We do have a language agnostic way to generate this kind of feature, as
implemented in #3213, but our experience with the "Python: Select
Interpreter" command highlights that making these language specific is
pretty helpful.

What this PR gets us for R (beyond the existing functionality) now is:

- Kick the R discovery process without reloading the whole app
- More control on how various R runtimes are presented (for example,
system R at the top of the list)

What we could add in the future includes:

- Show non-usable R installations

Related to the upcoming multiple sessions work and changes to the
interpreter dropdown, we'll need to add to the Positron API for
extensions so that extensions can say, "Here is my command to select an
interpreter", and we can use `r.selectInterpreter` together with
`python.setInterpreter`. cc @dhruvisompura


Let's run the @:console tests.

### Release Notes

#### New Features

- New "R: Select Interpreter" command to see all options for R
interpreters and choose one.

#### Bug Fixes

- N/A


### QA Notes

- Start off by uninstalling your release version of R and having some
older versions of R installed.
- Run Positron and note that the interpreter dropdown shows your
installed R versions.
- Run "R: Select Interpreter" and start/switch to different
interpreters.
- Leaving Positron open and running, go to the terminal and install R
from brew via `brew install r`.
- Note that this new version of R _does not appear_ in the interpreter
dropdown; Positron doesn't know about it.
- Run "R: Select Interpreter" and note that you do see the new homebrew
R:

![Screenshot 2025-01-31 at 2 57
19 PM](https://github.com/user-attachments/assets/e0e86f37-ce35-4040-b70f-c4afe095a3a1)

- Choose the new homebrew R and notice that it starts (we are
registering it on the fly).

When you are done, do `brew uninstall r` because you certainly do not
want that R brew installation hanging around. 😆

---------

Signed-off-by: Julia Silge <julia.silge@gmail.com>