Remove the Jupyter Adapter (#6242)
This change removes the Jupyter Adapter. It has been disabled by default
for several months, with the goal of keeping it available as a backup or
troubleshooting tool while the supervisor stabilizes. However, it's been
a long time since we had to ask anyone to enable it, so it's time to say
goodbye.

In addition to slimming down some code, this change also has the side
effect of speeding up the builds since the Jupyter Adapter consumes a
disproportionate amount of time in the module install step (during which
it is compiled with `node-gyp`).

Addresses #4579.

### QA Notes

As the Jupyter Adapter has been disabled by default for a while now, the
only part of this change visible to the user is that there's no longer a
_Kernel Supervisor > Enable_ setting.