e2e-test: add fixture to open folder (#6323)
### Summary
In the r-pkg-development test, there was an intermittent issue when
opening a folder. I determined that this was due to a lag in files
loading within the quick input dropdown.

To resolve this, I found the most reliable approach was to add each
directory in the path individually, ensuring the correct folder is
opened. I encapsulated this logic into a reusable fixture: `openFolder`.

### QA Notes

@:r-pkg-development @:test-explorer @:web