vscode-python upstream changes v2024.20.0 to v2025.0.0 (#6298)
Includes changes from:
[v2024.22.1](https://github.com/microsoft/vscode-python/releases/tag/v2024.22.1),
[v2024.22.2](https://github.com/microsoft/vscode-python/releases/tag/v2024.22.2),
and
[v2025.0.0](https://github.com/microsoft/vscode-python/releases/tag/v2025.0.0).

A lot of restructuring tests and removing now-unused code, especially
for tensorboard.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->

should pass all ci, lsp tasks should still work, etc