add option `python.interpreters.include`
This option allows user-specified Python search paths to be included as part of Python installation discovery. This is useful when the Python interpreter is not installed in a standard location.

The option is a list of unique paths to include in the search for Python interpreters. The paths can be absolute or relative to the user's home directory. The default value is an empty list.