Connections: Fix icon rendering on Windows (#6370)
This tries to fix an issue when rendering an icon on Windows. (It would
probably also fail on the Web version)
When loading a local asset, we have to convert it into a browserURI
using `FileAccess.uriToBrowserUri`, this wraps the URI with
`vscode-file://vscode-file/C://` which seems to be sufficient for the
file to be found when rendering.