Exclude ipykernel native libs for macOS universal release (#6377)
Fixing macOS releases. 

Skipping comparing files from the bundled ipykernel while preparing
macOS universal release.

To expedite fixing desktop macOS builds, it was decided to not try to
bundle ipykernel for both arm64 AND x86 macOS. Wasim updated the PR so
that we only bundle ipykernel for arm64 macOS for now. Intel x86
environments will need to continue install ipykernel on first use.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

Windows, Linux and MacOS arm64 should have ipykernel available in
Positron python sessions, where as macOS x86 will still need ipykernel
to be installed on first use.

---------

Signed-off-by: Wasim Lorgat <mwlorgat@gmail.com>
Co-authored-by: seem <mwlorgat@gmail.com>