Data Explorer: Compute summary statistics for pandas dtype=object columns having mixed types (#6391)
Addresses #3401. Currently, only the number of nulls and number of
unique values are own through an "other_stats" object that can capture
statistics we want to pass outside of the other types we support in the
summary pane.

There is no sparkline for now, but we could experiment with a frequency
table in follow up work.