Reticulate: Don't use bundled ipykernel (#6400)
With Reticulate, there's no way to control when the Python runtime will
be started. It's possible that the Python runtime is started and only
after some time it gets connected to Positron as a console. Thus there's
no way to reliably modify `PYTHONPATH` in order to use the bundled
Ipykernel.

---------

Co-authored-by: Chris Mead <chris.mead@posit.co>