Improves Main Interpreter Picker (#6363)
Adds more useable interface for primary interpreter picker.

Addresses #6150, #3596, #2143


https://github.com/user-attachments/assets/a9aa05ae-3c51-4a3b-bbec-5b0add8a9373


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Adds action to open Quick Pick to allow user to choose active runtime
session, or to create a new one
- Replaces main interpreter picker dropdown with button to trigger Quick
Pick above
- Modifies Positron action bar button to allow display of custom icon
(vs. only Codicons as was previously) to support language icons in new
interpreter button

#### Bug Fixes

- N/A


### QA Notes

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->

e2e: @:interpreter

---------

Signed-off-by: Sam Clark <samclark2015@users.noreply.github.com>
Co-authored-by: Dhruvi Sompura <dhruvi.sompura@posit.co>