Fix Multiple Console Sessions Feature Flag for Web (#6410)
### Release Notes
Addresses #6401 

The configuration scope for `console.multipleConsoleSessions` was set to
`MACHINE_OVERRIDABLE` which prevented enabling the feature flag as a
user setting in the Web build of Positron.

We do not require this scope so it has been removed and allows the
feature flag to be enabled which unblcosk CI testing for this new
feature!

#### New Features

- N/A

#### Bug Fixes

- N/A

### QA Notes
@:sessions @:web @:win

---------

Co-authored-by: Marie Idleman <marie.idleman@posit.co>